build: add flag to enable linking of common symbols
With this change, a new flag has been added to enable the linking of common symbols in the node build process. This is achieved by adding the -fcommon flag to the build script. The new flag should enable better compatibility with other builds and ensure a more robust production release.