feat: Add multithreading support for mining
This commit adds support for multithreading while mining a block. It introduces a new `bruteforce()` function and `bound` struct to divide the nonce range amongst the threads. The `mine()` function now accepts a `block` parameter instead of `block*` for improved safety. The `mining()` function has been modified to run multiple threads in parallel. The number of threads or thread count can be controlled by using the `-t` or `-thread` command line option. The commit message adheres to conventional commit format and uses an informal tone.