feat: Implement mining block feature
Create a genesis block, setting transactions to null and create a function to mine a new block. In the block chain function, the challenge checks if the current block hash meets the criteria of DIFFICULTY with the corresponding number of zeros. Mine function creates a transaction, increments the index number and attaches the previous hash to the new block. Added a function to get an empty address, get difficulty and changed public address size in the header file. For the node, added the function to report when a mined block occurs, used the "bruteforce" function to mine a block, creating a buffer node and broadcast it to all nodes.