feat: Add functions to count, populate and retrieve all keys in dictionary
Adds three new functions in dict.c file - count_keys, populate_keys and dict_get_all_keys - to count the number of keys in a dictionary, populate an array with all keys, and retrieve the array of keys respectively. These new functions can be used to handle the keys in a more programmatic way.