feat: Update function arguments to use pointers
This commit changes the `challenge` and `mine` functions in `blockchain.c` to use pointer arguments instead of value arguments. Additionally, the `hash_block` function is also updated to accept a pointer argument. In `node.c`, the `on_mined_block` function is updated to pass a chain pointer to the `challenge` function. This change is made to ensure that the functions can work properly with the correct data types. Finally, the `get_empty_address` function now uses `strncpy` to create a null-terminated public address string.